wip(db): port `vectorise` command to the new db layer
